refactor: LadderService -> ResultConverter 변경
