+++ src/main/java/nextstep/ladder/Application.java
@@ -3,14 +3,13 @@ package nextstep.ladder;
 import nextstep.ladder.controller.LadderController;
 import nextstep.ladder.generator.RandomGenerator;
 import nextstep.ladder.generator.RandomLadderGenerator;
-import nextstep.ladder.service.LadderService;
-import nextstep.ladder.service.LadderServiceImpl;
+import nextstep.ladder.converter.ResultConverter;
 
 public class Application {
 	public static void main(String[] args) {
 		RandomGenerator generator = new RandomLadderGenerator();
-		LadderService ladderService = new LadderServiceImpl();
-		LadderController ladderController = new LadderController(generator, ladderService);
+		ResultConverter resultConverter = new ResultConverter();
+		LadderController ladderController = new LadderController(generator, resultConverter);
 		ladderController.start();
 	}
 }
