+++ src/main/java/nextstep/ladder/controller/LadderController.java
@@ -1,22 +1,22 @@
 package nextstep.ladder.controller;
 
+import nextstep.ladder.converter.ResultConverter;
 import nextstep.ladder.domain.ExecutionResult;
 import nextstep.ladder.domain.Height;
 import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.Participants;
 import nextstep.ladder.dto.LadderResultDto;
 import nextstep.ladder.generator.RandomGenerator;
-import nextstep.ladder.service.LadderService;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
 
 public class LadderController {
 	private final RandomGenerator generator;
-	private final LadderService ladderService;
+	private final ResultConverter resultConverter;
 
-	public LadderController(RandomGenerator generator, LadderService ladderService) {
+	public LadderController(RandomGenerator generator, ResultConverter resultConverter) {
 		this.generator = generator;
-		this.ladderService = ladderService;
+		this.resultConverter = resultConverter;
 	}
 
 	public void start() {
@@ -26,7 +26,7 @@ public class LadderController {
 
 		Ladder ladder = Ladder.initialize(generator, participants.size(), height.getValue());
 
-		LadderResultDto resultDto = ladderService.calculateResult(ladder, participants, executionResult);
+		LadderResultDto resultDto = resultConverter.convert(ladder, participants, executionResult);
 
 		ResultView.printTitle();
 		ResultView.printParticipants(participants.getValues());
